requirements.txt
django
pyyaml
uritemplate
djangorestframework
djangorestframework-simplejwt
mysqlclient 
python-decouple
uritemplate
drf-spectacular
